Refactoring test to have permission to land the aircraft into the airport and creating the code to pass the test
