+++ lib/airport_details.rb
@@ -1,17 +1,10 @@
-#Class Weather
-class Weather
-	def initialize
-		@storm = false
-	end
-	def stormy?(status = true)
-		@storm = status
-	end
-end
-
+require 'weather'
 # Class to know the status of the aircraft and airport
-class Airport < Weather
-	def initialize()
+class Airport
+	attr_reader :weather
+	def initialize(weather = Weather.new)
 		@hangar = []
+		@weather = weather
 	end
 
 	DEFAULT_CAPACITY = 20
@@ -21,14 +14,14 @@ class Airport < Weather
 	end
 
 	def land
-		fail 'You can not land the aircraft because the airport is full' if full?
-		fail 'You can not land the aircraft because of the storm' if @storm == true
+		# fail 'You can not land the aircraft because the airport is full' if full?
+		# fail 'You can not land the aircraft because of the storm' if @weather.stormy?
 		@hangar << +1
-		@hangar.count
+		@hangar.count and true
 	end
 
 	def take_off
-		fail 'You can not take off the aircraft because of the storm' if @storm == true
+		fail 'You can not take off the aircraft because of the storm' if @weather.stormy?
 		@hangar.pop
 		@hangar
 	end
