+++ lib/weather.rb
@@ -1,0 +1,9 @@
+#Class Weather
+class Weather
+	def initialize
+		@storm = false
+	end
+	def stormy?(status = true)
+		@storm = status
+	end
+end
